refactor(utilities): utilities/ の格納場所を scripts/ に変更した
