+++ scripts/utilities/convertToLowerCasedArray.js
