+++ scripts/utilities/isEmail.js
