+++ scripts/utilities/isSelects.js
