+++ scripts/utilities/isStringOfNotEmpty.js
