+++ scripts/utilities/toBasename.js
